MAjor changes to box2d stuff
