+++ src/main/kotlin/eater/ecs/ashley/systems/Box2dDebugRenderSystem.kt
@@ -4,7 +4,6 @@ import com.badlogic.ashley.core.EntitySystem
 import com.badlogic.gdx.graphics.OrthographicCamera
 import com.badlogic.gdx.physics.box2d.Box2DDebugRenderer
 import com.badlogic.gdx.physics.box2d.World
-import eater.core.world
 
 
 class Box2dDebugRenderSystem(
@@ -15,18 +14,6 @@ class Box2dDebugRenderSystem(
 
     override fun update(deltaTime: Float) {
         debugRenderer.render(world, camera.combined)
-    }
-}
-
-class Box2dUpdateSystem(private val timeStep: Float, private val velIters: Int, private val posIters: Int): EntitySystem() {
-    var accumulator = 0f
-    override fun update(deltaTime: Float) {
-        val ourTime = deltaTime.coerceAtMost(timeStep * 2)
-        accumulator += ourTime
-        while (accumulator > timeStep) {
-            world().step(timeStep, velIters, posIters)
-            accumulator -= ourTime
-        }
     }
 }
