+++ src/main/kotlin/eater/ecs/ashley/systems/CameraFollowSystem.kt
@@ -4,8 +4,8 @@ package eater.ecs.ashley.systems
 import com.badlogic.ashley.core.Entity
 import com.badlogic.ashley.systems.IteratingSystem
 import com.badlogic.gdx.graphics.OrthographicCamera
-import eater.ecs.ashley.components.Box2d
 import eater.ecs.ashley.components.CameraFollow
+import eater.ecs.ashley.components.TransformComponent
 import ktx.ashley.allOf
 import ktx.math.vec2
 import ktx.math.vec3
@@ -17,7 +17,7 @@ open class CameraFollowSystem(
     IteratingSystem(
         allOf(
             CameraFollow::class,
-            Box2d::class
+            TransformComponent::class
         ).get()
     ) {
 
@@ -26,8 +26,8 @@ open class CameraFollowSystem(
 
     override fun processEntity(entity: Entity, deltaTime: Float) {
 
-        val body = Box2d.get(entity).body
-        cameraPosition.set(body.position)
+        val position = TransformComponent.get(entity).position
+        cameraPosition.set(position)
 
         camera.position.lerp(
             vec3(cameraPosition, 0f), alpha
